--- a/benches/lib.rs
+++ b/benches/lib.rs
@@ -1,7 +1,7 @@
 use std::hint::black_box;
 
 use criterion::{Criterion, criterion_group, criterion_main};
-use naijascript::arena::Arena;
+use naijascript::arena::{self, scratch_arena};
 use naijascript::syntax::parser::Parser;
 use naijascript::syntax::scanner::Lexer;
 
@@ -14,12 +14,16 @@ fn bench_skip_comment(c: &mut Criterion) {
         # Comment
         x add "baz"
     "#;
-
+    let arena = scratch_arena(None);
     c.bench_function("skip_comment", |b| {
         b.iter(|| {
-            let lexer = Lexer::new(src);
+            let offset = arena.offset();
+            let lexer = Lexer::new(src, &arena);
             let token: Vec<_> = lexer.collect();
             black_box(token);
+            unsafe {
+                arena.reset(offset);
+            }
         })
     });
 }
@@ -31,23 +35,27 @@ fn bench_scan_string(c: &mut Criterion) {
         # String escape sequence
         "abcd\nefgh\tijkl\"mnop\\qrst"
     "#;
+    let arena = scratch_arena(None);
     c.bench_function("scan_string", |b| {
         b.iter(|| {
-            let lexer = Lexer::new(src);
+            let offset = arena.offset();
+            let lexer = Lexer::new(src, &arena);
             let token: Vec<_> = lexer.collect();
             black_box(token);
+            unsafe {
+                arena.reset(offset);
+            }
         })
     });
 }
 
-fn bench_parser_basic_assignment(c: &mut Criterion) {
+fn bench_parse_assignment(c: &mut Criterion) {
     let src = "make x get 42";
-    let arena = Arena::new(4 * 1024).unwrap();
-
-    c.bench_function("parser_basic_assignment", |b| {
+    let arena = scratch_arena(None);
+    c.bench_function("parse_assignment", |b| {
         b.iter(|| {
             let offset = arena.offset();
-            let lexer = Lexer::new(src);
+            let lexer = Lexer::new(src, &arena);
             let mut parser = Parser::new(lexer, &arena);
             let (root, error) = parser.parse_program();
             black_box((root, error));
@@ -56,14 +64,13 @@ fn bench_parser_basic_assignment(c: &mut Criterion) {
     });
 }
 
-fn bench_parser_arithmetic_expression(c: &mut Criterion) {
+fn bench_parse_arithmetic_expr(c: &mut Criterion) {
     let src = "make result get 1 add 2 times 3 divide 4 minus 5";
-    let arena = Arena::new(4 * 1024).unwrap();
-
-    c.bench_function("parser_arithmetic_expression", |b| {
+    let arena = scratch_arena(None);
+    c.bench_function("parse_arithmetic_expression", |b| {
         b.iter(|| {
             let offset = arena.offset();
-            let lexer = Lexer::new(src);
+            let lexer = Lexer::new(src, &arena);
             let mut parser = Parser::new(lexer, &arena);
             let (root, error) = parser.parse_program();
             black_box((root, error));
@@ -72,7 +79,7 @@ fn bench_parser_arithmetic_expression(c: &mut Criterion) {
     });
 }
 
-fn bench_parser_function_definition(c: &mut Criterion) {
+fn bench_parse_fn_def(c: &mut Criterion) {
     let src = r#"
         do fibonacci(n) start
             if to say (n small pass 2) start
@@ -81,12 +88,11 @@ fn bench_parser_function_definition(c: &mut Criterion) {
             return fibonacci(n minus 1) add fibonacci(n minus 2)
         end
     "#;
-    let arena = Arena::new(4 * 1024).unwrap();
-
-    c.bench_function("parser_function_definition", |b| {
+    let arena = scratch_arena(None);
+    c.bench_function("parse_function_definition", |b| {
         b.iter(|| {
             let offset = arena.offset();
-            let lexer = Lexer::new(src);
+            let lexer = Lexer::new(src, &arena);
             let mut parser = Parser::new(lexer, &arena);
             let (root, error) = parser.parse_program();
             black_box((root, error));
@@ -95,7 +101,7 @@ fn bench_parser_function_definition(c: &mut Criterion) {
     });
 }
 
-fn bench_parser_control_flow(c: &mut Criterion) {
+fn bench_parse_ctrl_flow(c: &mut Criterion) {
     let src = r#"
         make i get 0
         jasi (i small pass 10) start
@@ -107,12 +113,11 @@ fn bench_parser_control_flow(c: &mut Criterion) {
             i get i add 1
         end
     "#;
-    let arena = Arena::new(4 * 1024).unwrap();
-
-    c.bench_function("parser_control_flow", |b| {
+    let arena = scratch_arena(None);
+    c.bench_function("parse_control_flow", |b| {
         b.iter(|| {
             let offset = arena.offset();
-            let lexer = Lexer::new(src);
+            let lexer = Lexer::new(src, &arena);
             let mut parser = Parser::new(lexer, &arena);
             let (root, error) = parser.parse_program();
             black_box((root, error));
@@ -121,9 +126,8 @@ fn bench_parser_control_flow(c: &mut Criterion) {
     });
 }
 
-fn bench_parser_large_program(c: &mut Criterion) {
+fn bench_parse_long_program(c: &mut Criterion) {
     let src = r#"
-        # Large program with multiple functions and complex expressions
         do factorial(n) start
             if to say (n small pass 2) start
                 return 1
@@ -163,12 +167,11 @@ fn bench_parser_large_program(c: &mut Criterion) {
             i get i add 1
         end
     "#;
-    let arena = Arena::new(4 * 1024).unwrap();
-
-    c.bench_function("parser_large_program", |b| {
+    let arena = scratch_arena(None);
+    c.bench_function("parse_long_program", |b| {
         b.iter(|| {
             let offset = arena.offset();
-            let lexer = Lexer::new(src);
+            let lexer = Lexer::new(src, &arena);
             let mut parser = Parser::new(lexer, &arena);
             let (root, error) = parser.parse_program();
             black_box((root, error));
@@ -177,18 +180,17 @@ fn bench_parser_large_program(c: &mut Criterion) {
     });
 }
 
-fn bench_parser_string_interpolation(c: &mut Criterion) {
+fn bench_parse_string_interpolation(c: &mut Criterion) {
     let src = r#"
         make name get "World"
         make greeting get "Hello {name}, how you dey?"
         shout(greeting)
     "#;
-    let arena = Arena::new(4 * 1024).unwrap();
-
-    c.bench_function("parser_string_interpolation", |b| {
+    let arena = scratch_arena(None);
+    c.bench_function("parse_string_interpolation", |b| {
         b.iter(|| {
             let offset = arena.offset();
-            let lexer = Lexer::new(src);
+            let lexer = Lexer::new(src, &arena);
             let mut parser = Parser::new(lexer, &arena);
             let (root, error) = parser.parse_program();
             black_box((root, error));
@@ -197,14 +199,14 @@ fn bench_parser_string_interpolation(c: &mut Criterion) {
     });
 }
 
-fn bench_parser_nested_expressions(c: &mut Criterion) {
+fn bench_parse_nested_expr(c: &mut Criterion) {
     let src = "make result get ((((1 add 2) times 3) divide 4) minus 5) add ((6 times 7) divide (8 add 9))";
-    let arena = Arena::new(4 * 1024).unwrap();
+    let arena = scratch_arena(None);
 
-    c.bench_function("parser_nested_expressions", |b| {
+    c.bench_function("parse_nested_expression", |b| {
         b.iter(|| {
             let offset = arena.offset();
-            let lexer = Lexer::new(src);
+            let lexer = Lexer::new(src, &arena);
             let mut parser = Parser::new(lexer, &arena);
             let (root, error) = parser.parse_program();
             black_box((root, error));
@@ -213,16 +215,19 @@ fn bench_parser_nested_expressions(c: &mut Criterion) {
     });
 }
 
-criterion_group!(
-    benches,
-    bench_skip_comment,
-    bench_scan_string,
-    bench_parser_basic_assignment,
-    bench_parser_arithmetic_expression,
-    bench_parser_function_definition,
-    bench_parser_control_flow,
-    bench_parser_large_program,
-    bench_parser_string_interpolation,
-    bench_parser_nested_expressions
-);
+fn bench(c: &mut Criterion) {
+    arena::init(128 * 1024 * 1024).unwrap();
+
+    bench_skip_comment(c);
+    bench_scan_string(c);
+    bench_parse_assignment(c);
+    bench_parse_arithmetic_expr(c);
+    bench_parse_fn_def(c);
+    bench_parse_ctrl_flow(c);
+    bench_parse_long_program(c);
+    bench_parse_string_interpolation(c);
+    bench_parse_nested_expr(c);
+}
+
+criterion_group!(benches, bench);
 criterion_main!(benches);