… .github/workflows/ci.yml
… Cargo.lock
… Cargo.toml
… src/arena/cow.rs
… src/arena/debug.rs
… src/arena/mod.rs
… src/arena/scratch.rs
… src/arena/string.rs
… src/bin/naija/main.rs
… src/bin/naijaup/main.rs
… src/diagnostics.rs
… src/resolver.rs
… src/runtime.rs
… src/simd/memchr2.rs
… src/simd/mod.rs
… src/syntax/parser.rs
… src/syntax/scanner.rs
… src/syntax/token.rs
… src/sys/mod.rs
… src/sys/unix.rs
… src/sys/wasm.rs
… src/sys/windows.rs
… tests/common.rs
… tests/parser.rs
… tests/resolver.rs
… tests/runtime.rs
… tests/scanner.rs