Regression bench result:

```shell
old_parser              time:   [2.5144 µs 2.5276 µs 2.5412 µs]
                        change: [−1.7067% −1.0911% −0.4782%] (p = 0.00 < 0.05)
                        Change within noise threshold.

arena_parser            time:   [1.0528 µs 1.0535 µs 1.0542 µs]
                        change: [−0.2727% +0.0250% +0.3310%] (p = 0.89 > 0.05)
                        No change in performance detected.
Found 8 outliers among 100 measurements (8.00%)
  3 (3.00%) high mild
  5 (5.00%) high severe
```