@xosnrdev 2025-08-22
Overall benchmark:

```shell
skip_comment            time:   [195.07 ns 196.38 ns 197.63 ns]
                        change: [−0.0199% +0.6734% +1.5479%] (p = 0.09 > 0.05)
                        No change in performance detected.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild

scan_string             time:   [107.08 ns 107.15 ns 107.23 ns]
                        change: [+1.0003% +1.4116% +1.9416%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 8 outliers among 100 measurements (8.00%)
  5 (5.00%) high mild
  3 (3.00%) high severe

parse_assignment        time:   [46.264 ns 46.358 ns 46.452 ns]
Found 6 outliers among 100 measurements (6.00%)
  4 (4.00%) high mild
  2 (2.00%) high severe

parse_arithmetic_expression
                        time:   [136.94 ns 137.07 ns 137.21 ns]
                        change: [+0.6667% +1.0046% +1.3419%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 12 outliers among 100 measurements (12.00%)
  7 (7.00%) high mild
  5 (5.00%) high severe

parse_function_definition
                        time:   [378.54 ns 379.02 ns 379.52 ns]
                        change: [−4.6307% −2.9479% −1.8686%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  3 (3.00%) high severe

parse_control_flow      time:   [485.94 ns 486.38 ns 486.92 ns]
                        change: [−15.864% −14.913% −14.157%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 11 outliers among 100 measurements (11.00%)
  4 (4.00%) high mild
  7 (7.00%) high severe

parse_long_program      time:   [1.1183 µs 1.1192 µs 1.1203 µs]
                        change: [−5.6228% −4.6228% −3.9159%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 10 outliers among 100 measurements (10.00%)
  4 (4.00%) high mild
  6 (6.00%) high severe

parse_string_interpolation
                        time:   [194.46 ns 194.57 ns 194.69 ns]
                        change: [−1.9205% −1.6445% −1.2168%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  1 (1.00%) high mild
  6 (6.00%) high severe

parse_nested_expression time:   [307.13 ns 307.71 ns 308.42 ns]
                        change: [−13.410% −13.132% −12.868%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 13 outliers among 100 measurements (13.00%)
  1 (1.00%) high mild
  12 (12.00%) high severe
```